pyproject.toml
[project]
name = "neural-condense-subnet"
version = "0.0.2"
description = "Neural Condense Subnet - Bittensor"
authors = [
    {name = "toilaluan"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "transformers==4.45.2",
    "bittensor==8.4.3",
    "accelerate==1.0.1",
    "datasets==2.21.0",
    "pre-commit==4.0.1",
    "peft==0.13.2",
    "sentencepiece==0.2.0",
    "protobuf==5.28.3",
    "flask==3.0.3",
    "gunicorn==23.0.0",
    "httpx==0.27.2",
    "wandb==0.18.6",
    "redis==5.2.0",
    "kvpress==0.0.2",
    "hf-transfer==0.1.8",
    "structlog==24.4.0",
    "minio==7.2.11",
    "tabulate==0.9.0",
    "deepspeed>=0.15.4",
    "torchmetrics>=1.6.0",
    "semantic-text-splitter>=0.19.0",
]

[project.urls]
repository = "https://github.com/condenses/neural-condense-subnet"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
